LibAV says it supports Opus... but doesn't.
Opus was not supported until after the distinguishing micro-version
numbering was introduced (see r2547), so it should be safe to test
on this.

If and when the LibAV folk fix this bug we will need to revisit this
patch to enable Opus for LibAV subsequent to the bug-fixing release.

Also see: http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=720563

git-svn-id: svn://daper.net/moc/trunk@2548 910807d9-36e0-0310-a014-e9ea483e2ba4