Remove WAV from ModPlug-supported formats.
make_modplug_data() allocates memory into which the entire sound file
is loaded.  For music module files, this is unlikely to be a problem.
However, for sampled formats where the size of the sound file is
proportional to the duration of the audio it is easy to trigger memory
exhaustion.

Amends: r2356

git-svn-id: svn://daper.net/moc/trunk@2554 910807d9-36e0-0310-a014-e9ea483e2ba4